Fixing a bug that was found in python when using Coil::computeAllMutualInductanceArrangements for a large number of calculations and large precision factor. The total number of issued threads was larger than the largest unsigned integer. Further monitoring may be necessary. So long as the total numPoints is lower than 2^31, or in other cases (numPoints * numConfiguration) / NTHREADS is lower than 2^31, everything should be fine.

+++ src/Coil/CUDAKernels/FieldA.cu
@@ -11,7 +11,7 @@ __global__
 void calculatePotentialSlow(long long opCount, CoilData coil, const VectorData *posArr, VectorData *resArr)
 {
     unsigned int index = threadIdx.x;
-    long long global_index = blockIdx.x * blockDim.x + index;
+    long long global_index = (long long) blockIdx.x * blockDim.x + index;
 
     if(global_index >= opCount)
         return;
@@ -166,7 +166,7 @@ void Calculate_hardware_accelerated_a (long long opCount, CoilData coil,
         printf("---------------------------------------------------\n");
     #endif
 
-    long long blocks = ceil(double(opCount) / NTHREADS);
+    int blocks = int(std::ceil(double(opCount) / NTHREADS));
 
     getBuffers(opCount);
